Associate cluster label with centroid
Summary
----
It seems to me that there was currently no way to get the centroid of a
cluster associated with its label. I need to know this for some external
usage.

This patch makes it so that the center now also returns the associated
cluster label making it easier to idenfity each cluster's centroid.